+++ src/leapfrogai_api/routers/openai/audio.py
@@ -3,9 +3,9 @@
 from itertools import chain
 from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException
-from leapfrogai_api.backends.grpc_client import create_transcription
-from leapfrogai_api.backends.helpers import read_chunks
-from leapfrogai_api.routers.openai.types import (
+from leapfrogai_api.backend.grpc_client import create_transcription
+from leapfrogai_api.backend.helpers import read_chunks
+from leapfrogai_api.backend.types import (
     CreateTranscriptionRequest,
     CreateTranscriptionResponse,
 )
