+++ src/leapfrogai_api/routers/openai/chat.py
@@ -2,14 +2,14 @@
 
 from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException
-from leapfrogai_api.backends.grpc_client import (
+from leapfrogai_api.backend.grpc_client import (
     chat_completion,
     stream_chat_completion,
 )
-from leapfrogai_api.backends.helpers import grpc_chat_role
+from leapfrogai_api.backend.helpers import grpc_chat_role
 from leapfrogai_api.utils import get_model_config
 from leapfrogai_api.utils.config import Config
-from leapfrogai_api.routers.openai.types import ChatCompletionRequest
+from leapfrogai_api.backend.types import ChatCompletionRequest
 import leapfrogai_sdk as lfai
