+++ src/leapfrogai_api/routers/openai/completions.py
@@ -2,8 +2,8 @@
 
 from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException
-from leapfrogai_api.backends.grpc_client import completion, stream_completion
-from leapfrogai_api.routers.openai.types import CompletionRequest
+from leapfrogai_api.backend.grpc_client import completion, stream_completion
+from leapfrogai_api.backend.types import CompletionRequest
 from leapfrogai_api.utils import get_model_config
 from leapfrogai_api.utils.config import Config
 import leapfrogai_sdk as lfai
